+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Change log
 ==========
 
+1.1.1
+-----
+
 * 🎁 Add "Update" / "Help" menu entries
 * 🎁 Add "Find ICAO airport codes" menu entries
 * 🎁 Add marker for unsaved changes